fix: meta-og api url 수정
