+++ backend/src/main/java/wooteco/prolog/article/ui/ArticleController.java
@@ -1,14 +1,20 @@
 package wooteco.prolog.article.ui;
 
+import java.net.URI;
+import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 import wooteco.prolog.article.application.ArticleService;
 import wooteco.prolog.login.domain.AuthMemberPrincipal;
 import wooteco.prolog.login.ui.LoginMember;
-
-import java.net.URI;
-import java.util.List;
 
 @RequiredArgsConstructor
 @RestController
@@ -43,11 +49,5 @@ public class ArticleController {
                                               @AuthMemberPrincipal final LoginMember member) {
         articleService.delete(id, member);
         return ResponseEntity.noContent().build();
-    }
-
-    @PostMapping("/parse-url")
-    public ResponseEntity<ArticleUrlResponse> parseUrl(@RequestBody final ArticleUrlRequest articleUrlRequest){
-        final ArticleUrlResponse response = articleService.parse(articleUrlRequest);
-        return ResponseEntity.ok(response);
     }
 }
