+++ backend/src/main/java/wooteco/prolog/common/exception/BadRequestCode.java
@@ -87,7 +87,8 @@ public enum BadRequestCode {
     ARTICLE_NOT_FOUND_EXCEPTION(12004, "ARTICLE_NOT_FOUND_EXCEPTION"),
     INVALID_ARTICLE_AUTHOR_EXCEPTION(12005, "INVALID_ARTICLE_AUTHOR_EXCEPTION"),
     ARTICLE_IMAGE_URL_NULL_OR_EMPTY_EXCEPTION(12006, "ARTICLE_IMAGE_URL_NULL_OR_EMPTY_EXCEPTION"),
-    ARTICLE_IMAGE_URL_OVER_LENGTH_EXCEPTION(12007, "ARTICLE_IMAGE_URL_OVER_LENGTH_EXCEPTION");
+    ARTICLE_IMAGE_URL_OVER_LENGTH_EXCEPTION(12007, "ARTICLE_IMAGE_URL_OVER_LENGTH_EXCEPTION"),
+    ARTICLE_INVALID_URL_EXCEPTION(12008, "ARTICLE_INVALID_URL_EXCEPTION");
 
     private int code;
     private String message;
